srt_to_product/pyproject.toml
[project]
name = "srt-to-product"
version = "0.1.0"
description = "AI-powered SRT to product video generation using DeepSeek API"
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "openai>=1.13.3", # DeepSeek API 兼容 OpenAI 接口
    "ffmpeg-python>=0.2.0", # 视频处理 (替代moviepy)
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pathlib>=1.0.1",
    "tenacity>=9.1.2", # 重试机制
    "pillow>=9.0.0", # 图像处理
    "imageio>=2.37.0", # 图像IO支持
    "imageio-ffmpeg>=0.6.0", # FFmpeg集成
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
exclude = [
    "input_srt",
    "output_videos", 
    "srt_product_env",
    "*.txt",
    "__pycache__",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"

[project.scripts]
srt-to-product = "run:main"
batch-srt-to-product = "batch_srt_to_product:main" 
